+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Features
 - Support `dotnet test` with `--filter` with properties `FullyQualifiedName` and `Name`. (Fix: #76)
 ### Console
-- Update `Process` to start using `UseShellExecute` true.
+- Update `ricaun.Revit.Installation` to `1.4.0` to fix Revit start with `UseShellExecute` to true.
 ### TestAdapter
 - Add `TestFilter` to filter `TestCase` using `FullyQualifiedName` and `Name`.
 - Update `AdapterLogger` to enable/disable logs.